requirements.txt
django-background-tasks==1.2.0
django==2.2.4
django-background-tasks==1.2.0